refactor: split commands.py into modular commands/ package
Restructures the monolithic commands.py (1094 lines) into a well-organized
package with separate modules by functionality:

- commands/__init__.py (110 lines) - Module exports, backward compatibility
- commands/base.py (196 lines) - Core registry, decorator, CommandContext
- commands/export.py (178 lines) - export, summarize, replay commands
- commands/llm.py (201 lines) - model, tools, context, tokens commands
- commands/meta.py (300 lines) - help, setup, plugin, impersonate commands
- commands/session.py (269 lines) - log, undo, edit, rename, fork, etc.

Benefits:
- Commands grouped by functionality (session, llm, export, meta)
- Easier maintenance and testing of individual command groups
- Clearer dependencies with explicit imports
- Extensible design for adding new command modules

Technical changes:
- Used TYPE_CHECKING for forward references to avoid circular imports
- Used lazy imports inside functions for heavy dependencies
- Added llm pre-import to resolve pre-existing config<->llm circular dependency

All existing tests pass (1049 passed).
